Implementata la possibilità di avere Preprocessor differenti tramite la classe astratta Preprocessor.py
